arreglo tabla respuesta

+++ src/main/java/com/example/demo/repository/RespuestaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RespuestaRepository extends JpaRepository<Respuesta, Long> {
+public interface RespuestaRepository extends JpaRepository<Respuesta, Integer> {
 }
